+++ src/ladb_opencutlist/ruby/utils/model_utils.rb
@@ -9,9 +9,30 @@ module Ladb::OpenCutList
         model.entities.each { |entity|
           if entity.entityID == entity_id
             return entity
+          elsif entity.is_a? Sketchup::Group
+            e = find_entity_into_by_id(entity, entity_id)
+            return e unless e.nil?
+          elsif entity.is_a? Sketchup::ComponentInstance
+            e = find_entity_into_by_id(entity.definition, entity_id)
+            return e unless e.nil?
           end
         }
       end
+      nil
+    end
+
+    def self.find_entity_into_by_id(definition_or_group, entity_id)
+      definition_or_group.entities.each { |entity|
+        if entity.entityID == entity_id
+          return entity
+        elsif entity.is_a? Sketchup::Group
+          e = find_entity_into_by_id(entity, entity_id)
+          return e unless e.nil?
+        elsif entity.is_a? Sketchup::ComponentInstance
+          e = find_entity_into_by_id(entity.definition, entity_id)
+          return e unless e.nil?
+        end
+      }
       nil
     end